polyfill for Storage
When developer wants to check if localStorage and sessionStorage are available in current environment he often does the following ([source](https://www.w3schools.com/html/html5_webstorage.asp):
```
if (typeof(Storage) !== "undefined") {
    // Code for localStorage/sessionStorage.
} else {
    // Sorry! No Web Storage support..
}
```
So I propose to also polyfill the `Storage` function in this library. Thanks in advance for considering it.